libtrx/text: move Text_GetHeight to libtrx
