+++ src/tr1/game/text.h
@@ -11,5 +11,4 @@
 #define TEXT_OUTLINE_THICKNESS 2
 
 RGBA_8888 Text_GetMenuColor(MENU_COLOR color);
-int32_t Text_GetHeight(const TEXTSTRING *text);
 void Text_DrawText(TEXTSTRING *text);
