+++ src/tr2/game/text.h
@@ -6,8 +6,6 @@
 
 #define TEXT_HEIGHT 15
 
-int32_t __cdecl Text_GetHeight(const TEXTSTRING *text);
-
 void __cdecl Text_DrawBorder(
     int32_t x, int32_t y, int32_t z, int32_t width, int32_t height);
 void __cdecl Text_DrawText(TEXTSTRING *text);
